+++ src/Lykke.Service.HFT.Core/Domain/Accuracy.cs
@@ -1,17 +1,17 @@
 ﻿using Common;
-using Lykke.Service.Assets.Client.Models;
+using Lykke.Service.Assets.Client.Models.v3;
 
 namespace Lykke.Service.HFT.Core.Domain
 {
     public static class Accuracy
     {
-        public static double TruncatePrice(this double price, AssetPair assetPair, bool inverted = false)
+        public static decimal TruncatePrice(this decimal price, AssetPair assetPair, bool inverted = false)
             => price.TruncateDecimalPlaces(inverted ? assetPair.InvertedAccuracy : assetPair.Accuracy);
 
-        public static double? TruncatePrice(this double? price, AssetPair assetPair, bool inverted = false)
+        public static decimal? TruncatePrice(this decimal? price, AssetPair assetPair, bool inverted = false)
             => price?.TruncateDecimalPlaces(inverted ? assetPair.InvertedAccuracy : assetPair.Accuracy);
 
-        public static double TruncateVolume(this double price, Asset asset)
+        public static decimal TruncateVolume(this decimal price, Asset asset)
             => price.TruncateDecimalPlaces(asset.Accuracy);
     }
 }
